app/pipeline/ingest/py/requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
feedparser==6.0.11
lxml==5.1.0
pika==1.3.2
python-dateutil==2.8.2
PyYAML==6.0.1
sgmllib3k==1.0.0
soupsieve==2.5
six==1.16.0
xxhash==3.4.1
requests==2.31.0